+++ algoritmo4.py
@@ -1,8 +1,8 @@
 #Algoritmo que receba dois números e calcule a subtração entre eles.
-#Lê o primeiro número e converte para inteiro
-N1 = int(input("Informe um número: "))
-#Lê o segundo número e converte para inteiro
-N2 = int(input("Informe outro número: "))
+#Lê o primeiro número
+N1 = float(input("Informe um número: "))
+#Lê o segundo número
+N2 = float(input("Informe outro número: "))
 #Calcule a subtração entre os dois números
 R = N1 - N2
 #Mostre o resultado da subtração
